pyproject.toml
[project]
name = "my-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.135.1",
    "google-genai>=1.67.0",
    "python-dotenv>=1.2.2",
    "streamlit>=1.55.0",
]

[dependency-groups]
dev = [
    "black>=26.3.1",
    "pytest>=9.0.2",
    "ruff>=0.15.6",
]
